requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests==2.24.0
pyodbc==4.0.30
azure-identity==1.4.1
azure-keyvault-secrets==4.2.0